LayoutTests/inspector/network/clear-resource-data-on-navigate-expected.txt

Tests for Network.setClearResourceDataOnNavigate.


== Running test suite: Network.setClearResourceDataOnNavigate
-- Running test case: Network.setClearResourceDataOnNavigate.DefaultBehavior
Loading subresource...
PASS: Should be able to get response body before disable.
Disabling and enabling Network agent...
Requesting body for old resource...
PASS: Should not be able to get response body after disable/enable cycle.

-- Running test case: Network.setClearResourceDataOnNavigate.Disabled
Loading subresource...
PASS: Should be able to get response body before disable.
Disabling and enabling Network agent...
Requesting body for old resource...
PASS: Should not be able to get response body after disable/enable cycle.

-- Running test case: Network.setClearResourceDataOnNavigate.ToggledBackOn
Loading subresource...
Setting clearResourceDataOnNavigate back to true...
Disabling and enabling Network agent...
Requesting body for old resource...
PASS: Should not be able to get response body after disable/enable cycle.
